requirements.txt
streamlit
langchain-google-genai
langchain-community
langchain
chromadb
docling
python-dotenv
pysqlite3-binary
opencv-python-headless
google-generativeai>=0.8.0
langchain-chroma
langchain-huggingface
python-docx
